refactor: extract auto-memory chunker to scripts/lib (no behaviour change)
The session-end, pre-compact, and stop hooks all carried near-duplicate
copies of the salience scoring + content extraction code (~250 lines
each). Tightening rules has to land in one place, not three.

This commit consolidates them into scripts/lib/extract-memorable-segments.mjs
with knobs that preserve each hook's existing behaviour:

- session-end keeps DEFAULT_CATEGORY_THRESHOLDS, MAX_AUTO_MEMORIES=5,
  hookTag='session-end', frequency boost on
- pre-compact keeps PRE_COMPACT_CATEGORY_THRESHOLDS (raised +0.10 in
  v4.11.0), MAX_AUTO_MEMORIES=2
- stop-hook keeps the lighter 'stop' extractor mode (no architecture /
  important-note), PRE_COMPACT_CATEGORY_THRESHOLDS, no frequency boost,
  hookTag='source:stop-hook'

The keyword arrays + detect helpers are also exported so stop-hook's
isSalientWindow probe can keep using them without re-declaring.

Behaviour is preserved exactly. The three Phase-1 regression tests
remain red — they only turn green after Phase 3 (defence pipeline
wire-up) and Phase 5 (chunker tightening). All 87 other test suites
pass.
